rename about sections
