+++ client/src/components/about/AboutHeroSection.jsx
