+++ client/src/components/about/index.js
@@ -1,5 +1,5 @@
-export { default as AboutSectionOne } from './AboutSectionOne'
-export { default as AboutSectionTwo } from './AboutSectionTwo'
-export { default as AboutSectionThree } from './AboutSectionThree'
-export { default as AboutSectionFour } from './AboutSectionFour'
-export { default as AboutSectionFive } from './AboutSectionFive'
+export { default as AboutIntroSection } from './AboutIntroSection'
+export { default as AboutHeroSection } from './AboutHeroSection'
+export { default as AboutSkillsSection } from './AboutSkillsSection'
+export { default as AboutBackgroundSection } from './AboutBackgroundSection'
+export { default as AboutProjectsSection } from './AboutProjectsSection'
