+++ client/src/pages/About.jsx
@@ -1,9 +1,9 @@
 import {
-  AboutSectionOne,
-  AboutSectionTwo,
-  AboutSectionThree,
-  AboutSectionFour,
-  AboutSectionFive,
+  AboutIntroSection,
+  AboutHeroSection,
+  AboutSkillsSection,
+  AboutBackgroundSection,
+  AboutProjectsSection,
 } from '../components/about'
 import { motion } from 'framer-motion'
 
@@ -13,11 +13,11 @@ const About = () => (
     animate={{ opacity: 1 }}
     exit={{ opacity: 0, transition: { duration: 0.05 } }}
   >
-    <AboutSectionOne />
-    <AboutSectionTwo />
-    <AboutSectionThree />
-    <AboutSectionFour />
-    <AboutSectionFive />
+    <AboutIntroSection />
+    <AboutHeroSection />
+    <AboutSkillsSection />
+    <AboutBackgroundSection />
+    <AboutProjectsSection />
   </motion.div>
 )
